# docs/swagger/swagger.yaml
components:
  schemas:
    api.CheckpointStatus:
      properties:
        dataset:
          type: string
        duration_seconds:
          type: number
        error:
          type: string
        row_count:
          type: integer
        status:
          type: string
        table:
          type: string
      type: object
    api.LoadStatusResponse:
      properties:
        checkpoints:
          items:
            $ref: '#/components/schemas/api.CheckpointStatus'
          type: array
          uniqueItems: false
        load_id:
          type: string
        started_at:
          type: string
        status:
          type: string
      type: object
    api.OpenFDAActiveIngredient:
      properties:
        name:
          type: string
        strength:
          type: string
      type: object
    api.OpenFDAError:
      properties:
        error:
          $ref: '#/components/schemas/api.OpenFDAErrorDetail'
      type: object
    api.OpenFDAErrorDetail:
      properties:
        code:
          type: string
        message:
          type: string
      type: object
    api.OpenFDAMeta:
      properties:
        disclaimer:
          type: string
        last_updated:
          type: string
        license:
          type: string
        results:
          $ref: '#/components/schemas/api.OpenFDAPagination'
        terms:
          type: string
      type: object
    api.OpenFDANested:
      properties:
        is_original_packager:
          items:
            type: boolean
          type: array
          uniqueItems: false
        manufacturer_name:
          items:
            type: string
          type: array
          uniqueItems: false
        rxcui:
          items:
            type: string
          type: array
          uniqueItems: false
        spl_set_id:
          items:
            type: string
          type: array
          uniqueItems: false
        unii:
          items:
            type: string
          type: array
          uniqueItems: false
        upc:
          items:
            type: string
          type: array
          uniqueItems: false
      type: object
    api.OpenFDAPackaging:
      properties:
        description:
          type: string
        marketing_start_date:
          type: string
        package_ndc:
          type: string
        sample:
          type: boolean
      type: object
    api.OpenFDAPagination:
      properties:
        limit:
          type: integer
        skip:
          type: integer
        total:
          type: integer
      type: object
    api.OpenFDAProduct:
      properties:
        active_ingredients:
          items:
            $ref: '#/components/schemas/api.OpenFDAActiveIngredient'
          type: array
          uniqueItems: false
        application_number:
          type: string
        brand_name:
          type: string
        brand_name_base:
          type: string
        dosage_form:
          type: string
        finished:
          type: boolean
        generic_name:
          type: string
        labeler_name:
          type: string
        marketing_category:
          type: string
        marketing_start_date:
          type: string
        openfda:
          $ref: '#/components/schemas/api.OpenFDANested'
        packaging:
          items:
            $ref: '#/components/schemas/api.OpenFDAPackaging'
          type: array
          uniqueItems: false
        pharm_class:
          items:
            type: string
          type: array
          uniqueItems: false
        product_id:
          type: string
        product_ndc:
          type: string
        product_type:
          type: string
        route:
          items:
            type: string
          type: array
          uniqueItems: false
        spl_id:
          type: string
      type: object
    api.OpenFDAResponse:
      properties:
        meta:
          $ref: '#/components/schemas/api.OpenFDAMeta'
        results:
          items:
            $ref: '#/components/schemas/api.OpenFDAProduct'
          type: array
          uniqueItems: false
      type: object
    api.TriggerLoadRequest:
      properties:
        datasets:
          items:
            type: string
          type: array
          uniqueItems: false
        force:
          type: boolean
      type: object
    api.TriggerLoadResponse:
      properties:
        datasets:
          items:
            type: string
          type: array
          uniqueItems: false
        load_id:
          type: string
        status:
          type: string
      type: object
    store.PackageResult:
      properties:
        description:
          type: string
        ndc:
          type: string
        sample:
          type: boolean
      type: object
    store.ProductResult:
      properties:
        active_ingredients:
          type: string
        application_number:
          type: string
        brand_name:
          type: string
        dea_schedule:
          type: string
        dosage_form:
          type: string
        generic_name:
          type: string
        manufacturer:
          type: string
        marketing_category:
          type: string
        matched_package:
          type: string
        packages:
          items:
            $ref: '#/components/schemas/store.PackageResult'
          type: array
          uniqueItems: false
        pharm_classes:
          type: string
        product_ndc:
          type: string
        route:
          type: string
        strength:
          type: string
        strength_unit:
          type: string
      type: object
    store.StatsResult:
      properties:
        applications:
          type: integer
        last_loaded:
          type: string
        load_duration_seconds:
          type: number
        packages:
          type: integer
        products:
          type: integer
      type: object
  securitySchemes:
    ApiKeyAuth:
      in: header
      name: X-API-Key
      type: apiKey
externalDocs:
  description: ""
  url: ""
info:
  description: FDA NDC Directory bulk loader and REST API. Ingests NDC Directory and
    Drugs@FDA datasets daily, serves via REST with full-text search and NDC format
    normalization.
  title: ndc-loader API
  version: "1.0"
openapi: 3.1.0
paths:
  /api/admin/load:
    post:
      description: Start a manual FDA data load. Returns 409 if a load is already
        in progress.
      requestBody:
        content:
          application/json:
            schema:
              oneOf:
              - type: object
              - $ref: '#/components/schemas/api.TriggerLoadRequest'
                description: Load options
                summary: body
        description: Load options
      responses:
        "202":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.TriggerLoadResponse'
          description: Accepted
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized
        "409":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Conflict
      security:
      - ApiKeyAuth: []
      summary: Trigger a data load
      tags:
      - Admin
  /api/admin/load/{loadID}:
    get:
      description: Returns checkpoint progress for a specific load operation.
      parameters:
      - description: Load ID (UUID)
        in: path
        name: loadID
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.LoadStatusResponse'
          description: OK
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
      security:
      - ApiKeyAuth: []
      summary: Check load status
      tags:
      - Admin
  /api/ndc/{ndc}:
    get:
      description: Accepts any NDC format (hyphenated, unhyphenated, 2-segment, 3-segment).
        Returns product details with packages.
      parameters:
      - description: NDC code (e.g. 0002-1433, 0002-1433-61, 00021433)
        in: path
        name: ndc
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/store.ProductResult'
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Not Found
      security:
      - ApiKeyAuth: []
      summary: Look up a product by NDC code
      tags:
      - Query
  /api/ndc/{ndc}/packages:
    get:
      description: Returns all package configurations for a given product NDC.
      parameters:
      - description: Product NDC (e.g. 0002-1433)
        in: path
        name: ndc
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized
      security:
      - ApiKeyAuth: []
      summary: List packages for a product NDC
      tags:
      - Query
  /api/ndc/search:
    get:
      description: Full-text search across brand names, generic names, and manufacturers.
        Supports prefix matching.
      parameters:
      - description: Search query (e.g. metformin)
        in: query
        name: q
        required: true
        schema:
          type: string
      - description: Results per page
        in: query
        name: limit
        schema:
          default: 50
          type: integer
      - description: Pagination offset
        in: query
        name: offset
        schema:
          default: 0
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: OK
        "400":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized
      security:
      - ApiKeyAuth: []
      summary: Search drugs by name
      tags:
      - Query
  /api/ndc/stats:
    get:
      description: Returns row counts, last load time, and data freshness.
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/store.StatsResult'
          description: OK
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Internal Server Error
      security:
      - ApiKeyAuth: []
      summary: Dataset statistics
      tags:
      - Query
  /api/openfda/ndc.json:
    get:
      description: Drop-in replacement for the openFDA /drug/ndc.json endpoint. Returns
        identical response format. Supports openFDA search syntax (field:value, exact
        phrases, AND via +).
      parameters:
      - description: openFDA search query (e.g. brand_name:metformin, product_ndc:\
        in: query
        name: search
        required: true
        schema:
          type: string
      - description: Results per page (default 1, max 1000)
        in: query
        name: limit
        schema:
          default: 1
          type: integer
      - description: Pagination offset
        in: query
        name: skip
        schema:
          default: 0
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.OpenFDAResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.OpenFDAError'
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/api.OpenFDAError'
          description: Not Found
      security:
      - ApiKeyAuth: []
      summary: openFDA-compatible NDC search
      tags:
      - OpenFDA
  /health:
    get:
      description: Returns service health status and data freshness. Degrades to "degraded"
        when data is >48 hours old. No authentication required.
      responses:
        "200":
          content:
            application/json:
              schema:
                additionalProperties: {}
                type: object
          description: OK
      summary: Health check
      tags:
      - Operations
servers:
- url: localhost:8081/